refactor(ESP32): Redefine i2c_deinit behaviour
Applying the same logic as with the i2c_init method, we should always free the
I2C interface if the I2C bus is not initialized anymore.

Furthermore, it should be possible to decide if we also want to de-initialize
the I2C bus.

This commit also added a bool to i2c_deinit to de-initialize the I2C bus if
true.